(Class 3) Created spawners and pacer scripts
Created multiple scripts:

Spawner: When left click is pressed, spawns in a prefab that destroys itself in the order it was spawned in at (e.g. prefab 11 would destroy after 1.1 seconds, prefab 64 for destroy after 6.4 seconds). When right mouse button is pressed, it destroys all prefabs spawned in.

Timed spawner: Creates a prefab every 3 seconds

Pacer: Similar to "squareMover" with better logic, once an object reaches one side of the screen, it switches directions to another.